requirements.txt
Flask==0.10.1
pyrebase==0.2.932
gunicorn==19.2.1
PyJWT==1.3.0
tweepy==3.3.0